restructure repo to be simpler
