+++ .gitignore
@@ -1,2 +1,3 @@
 .go
 _output
+velero-plugin-example
